pyproject.toml
[tool.poetry]
name = "projeto-streamlit-aws"
version = "0.1.0"
description = ""
authors = ["Mairon.az <duartemairon@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
